+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* YoastSEO: do not copy title and metadesc
+
 ### 1.0.8: August 29th, 2018
 * Move SmushIt to submenu
 * ACF Disable Autocomplete